+++ src/main/java/exercise2/beans/Calculator.java
@@ -1,5 +1,5 @@
 package exercise2.beans;
 
 public interface Calculator {
-    int calculate(int input1, int input2, String operator);
+    double calculate(int input1, int input2, String operator);
 }
